new_interceptor_examples/code_injection/pyproject.toml
[tool.poetry]
name = "qiskit-interceptor-example"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
qiskit = "^0.26.0"
dwave-ocean-sdk = "^3.3.0"
amazon-braket-sdk = "^1.6.2"
amazon-braket-ocean-plugi = "^1.0.3.post1"
click = "^8.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
